✨ feat: adicionar componentes de controle de votação e gerenciamento de jogadores
